% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan_updated.R
\name{MiXcan_association_sep}
\alias{MiXcan_association_sep}
\title{Cell-Type-Aware Association (Separate Models)}
\usage{
MiXcan_association_sep(new_y, new_cov, new_outcome, family = "gaussian")
}
\arguments{
\item{new_y}{A matrix or data frame of predicted expression for two cell types. The first column will be used for cell 1, the second for cell 2.}

\item{new_cov}{A data frame of covariates (e.g., age, PCs).}

\item{new_outcome}{A vector or one-column matrix of phenotype values (e.g., binary case/control or continuous trait).}

\item{family}{A GLM family, either \code{"gaussian"} or \code{"binomial"}.}
}
\value{
A data frame with estimates, standard errors, and p-values for each cell type, and a combined p-value using ACAT.
}
\description{
This function performs two separate generalized linear models (GLMs) for testing the association between each cell-type-specific predicted expression
and the outcome, adjusting for covariates. It uses ACAT to combine p-values from both models.
}
