% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_estimation.R
\name{pi_estimate}
\alias{pi_estimate}
\title{Estimate cell-type proportions (SMiXcan wrapper)}
\usage{
pi_estimate(...)
}
\arguments{
\item{...}{Arguments passed to MiXcan's internal function.}
}
\value{
The value returned by MiXcan's `pi_estimate`.
}
\description{
This is a thin wrapper that calls MiXcan's internal `pi_estimate`
(if available). You must have MiXcan installed.
}
\seealso{
\code{MiXcan} package.
}
