% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_prediction_model.R
\name{train_prediction_model}
\alias{train_prediction_model}
\title{Train Cell-Type-Specific Prediction Model Using MiXcan}
\usage{
train_prediction_model(
  y.train,
  x.train,
  pi.train,
  cov = NULL,
  xNameMatrix = NULL,
  yName = NULL,
  foldid = NULL
)
}
\arguments{
\item{y.train}{Gene expression levels in training data.}

\item{x.train}{Genotype matrix in training data.}

\item{pi.train}{Estimated cell-type proportions in training data.}

\item{cov}{A N by Q matrix for the covariates adjusted in the model (e.g. age, population stratification).}

\item{xNameMatrix}{Default is NULL. A matrix to save theX matrix information,
such as variable ID, position, rsid, ref_allele, eff_allele.}

\item{yName}{Default is NULL. A row vector to save the expression information, such as gene ID, gene name.}

\item{foldid}{Default is NULL. 10-fold cross-validation (CV) is used in our pipeline. A random split}
}
\value{
A list containing:
\describe{
  \item{W1}{Weight vector for cell type 1}
  \item{W2}{Weight vector for cell type 2}
  \item{selected_snp}{Indices of selected SNPs}
  \item{MiXcan_weight_result}{Total result matrix}
}
}
\description{
Train Cell-Type-Specific Prediction Model
}
\details{
Trains a cell-type-aware prediction model using MiXcan, estimating weights for SNPs across cell types.
}
