% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_prediction_model.R
\name{train_prediction_model}
\alias{train_prediction_model}
\title{Train Cell-Type-Specific Prediction Model Using MiXcan}
\usage{
train_prediction_model(y.train, x.train, pi.train)
}
\arguments{
\item{y.train}{Gene expression levels in training data.}

\item{x.train}{Genotype matrix in training data.}

\item{pi.train}{Estimated cell-type proportions in training data.}
}
\value{
A list containing:
\describe{
  \item{W1}{Weight vector for cell type 1}
  \item{W2}{Weight vector for cell type 2}
  \item{selected_snp}{Indices of selected SNPs}
}
}
\description{
Train Cell-Type-Specific Prediction Model
}
\details{
Trains a cell-type-aware prediction model using MiXcan, estimating weights for SNPs across cell types.
}
